refactor(web): extract App.vue composables for page title, auth, sidebar, detail panel and file preview
