+++ .changeset/web-storage-appearance-notification.md
@@ -2,4 +2,4 @@
 "@moonshot-ai/kimi-code": patch
 ---
 
-Consolidate web client localStorage access and decouple appearance/notification state into dedicated modules.
+Consolidate web client localStorage access and split the root state store and app shell into focused composables.
